- removed old epxlorer and renamed explorer2 to explorer - added extension registering to the explorer - abstractation for the FlowGraph writing (as exporation extension)
