+++ dnWalker/Concolic/XmlExplorationExporter.cs
@@ -173,9 +173,9 @@ namespace dnWalker.Concolic
     }
 
 
-    public class XmlExplorationExporter : IExplorationExtension, IDisposable
+    public class XmlExplorationExporter : IExplorationExtension
     {
-        private Explorer2 _explorer;
+        private Explorer _explorer;
         private readonly string _file;
 
         private readonly XElement _rootElement;
@@ -190,7 +190,7 @@ namespace dnWalker.Concolic
             _rootElement = new XElement("Explorations");
         }
 
-        public void Register(Explorer2 explorer)
+        public void Register(Explorer explorer)
         {
             _explorer = explorer;
             _explorer.ExplorationStarted += OnExplorationStarted;
@@ -201,7 +201,7 @@ namespace dnWalker.Concolic
             _explorer.IterationFinished += OnIterationFinished;
         }
 
-        public void Dispose()
+        public void Unregister(Explorer explorer)
         {
             _explorer.ExplorationStarted -= OnExplorationStarted;
             _explorer.ExplorationFinished -= OnExplorationFinished;
@@ -210,6 +210,7 @@ namespace dnWalker.Concolic
             _explorer.IterationStarted -= OnIterationStarted;
             _explorer.IterationFinished -= OnIterationFinished;
         }
+
 
         private void SaveData()
         {
@@ -222,9 +223,9 @@ namespace dnWalker.Concolic
             _currentExplorationElement = new XElement("Exploration");
             _currentExplorationElement.SetAttributeValue("AssemblyName", e.AssemblyName);
             _currentExplorationElement.SetAttributeValue("AssemblyFileName", e.AssemblyFileName);
-            _currentExplorationElement.SetAttributeValue("MethodNameSignature", e.MethodName);
+            _currentExplorationElement.SetAttributeValue("MethodNameSignature", e.MethodSignature);
             _currentExplorationElement.SetAttributeValue("IsStatic", e.IsStatic);
-            _currentExplorationElement.SetAttributeValue("Solver", e.Solver);
+            _currentExplorationElement.SetAttributeValue("Solver", e.SolverType.FullName);
 
 
             _rootElement.Add(_currentExplorationElement);
@@ -255,6 +256,16 @@ namespace dnWalker.Concolic
         private void OnIterationFinished(object sender, IterationFinishedEventArgs e)
         {
             SaveData();
+        }
+    }
+
+    public static class XmlExplorationExporterExtensions
+    {
+        public static XmlExplorationExporter ExportXmlData(this Explorer explorer, string outputFile = "data.xml")
+        {
+            XmlExplorationExporter exporter = new XmlExplorationExporter(outputFile);
+            explorer.AddExtension(exporter);
+            return exporter;
         }
     }
 }
